+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This file lists the changes that have occurred since January 2024 in the project
 * Rename `--latency-peer-server` to `--latency-peer-address`
 * Continuous clock synchronization with the latency monitor
 * Support opening result files in the GUI by drag and drop
+* Add `--out-name` command line option to specify result filename prefix
+* Change filename prefix for both raw result and plots to `test`
 
 ## 0.2 - 2024-08-29